# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Data Smith Api Documentation",
        "description": "Data Smith  Api Documentation",
        "contact": {
            "name": "Lee Dong Kwan",
            "email": "dklee@lennon.co.kr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/part"
        }
    ],
    "paths": {
        "/list": {
            "get": {
                "tags": [
                    "- 전체 리스트 요청"
                ],
                "summary": "",
                "description": "전체 리스트 요청",
                "operationId": "4519c82ab6f988d56bb33a3f6c4acb2d",
                "responses": {
                    "200": {
                        "description": "결과값",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result_code": {
                                            "description": "성공:0, 실패:-1",
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "result_message": {
                                            "description": "성공:Success, 실패:에러메세지",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "result_data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "이름",
                                                        "type": "string",
                                                        "example": "이름 1"
                                                    },
                                                    "description": {
                                                        "description": "설명",
                                                        "type": "string",
                                                        "example": "설명 1"
                                                    },
                                                    "data_json": {
                                                        "description": "JSON 문자열",
                                                        "type": "string",
                                                        "example": "{}"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/list/{id}": {
            "get": {
                "tags": [
                    "- 지정 파츠 요청"
                ],
                "summary": "",
                "description": "지정 파츠 요청",
                "operationId": "e5372e443cea6e800b17e6f943e3bfa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "지정 파츠 id",
                        "required": true,
                        "schema": {
                            "type": "int"
                        },
                        "examples": {
                            "int": {
                                "summary": "paramter",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "결과값",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result_code": {
                                            "description": "성공:0, 실패:-1",
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "result_message": {
                                            "description": "성공:Success, 실패: 파츠가 존재 하지 않음",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "result_data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "이름",
                                                        "type": "string",
                                                        "example": "이름 1"
                                                    },
                                                    "description": {
                                                        "description": "설명",
                                                        "type": "string",
                                                        "example": "설명 1"
                                                    },
                                                    "data_json": {
                                                        "description": "JSON 문자열",
                                                        "type": "string",
                                                        "example": "{}"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/listByName/{name}": {
            "get": {
                "tags": [
                    "- 지정 파츠 요청(name 지정)"
                ],
                "summary": "",
                "description": "지정 파츠 요청",
                "operationId": "a30f7f66017dd52a34856bd7544d80b9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "지정 파츠 name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "paramter",
                                "value": "name_01"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "결과값",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result_code": {
                                            "description": "성공:0, 실패:-1",
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "result_message": {
                                            "description": "성공:Success, 실패:파츠가 존재 하지 않음",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "result_data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "이름",
                                                        "type": "string",
                                                        "example": "이름 1"
                                                    },
                                                    "description": {
                                                        "description": "설명",
                                                        "type": "string",
                                                        "example": "설명 1"
                                                    },
                                                    "data_json": {
                                                        "description": "JSON 문자열",
                                                        "type": "string",
                                                        "example": "{}"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        },
        "/save": {
            "post": {
                "tags": [
                    "- 파츠 저장"
                ],
                "summary": "",
                "description": "파츠 저장, 지정 이름, 설명을 json data로 요청하면 저장한다",
                "operationId": "83982f1c8038ef0325623908c7f4c4d6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "이름 1",
                                        "type": "string",
                                        "example": "이름 1"
                                    },
                                    "description": {
                                        "description": "설명 1",
                                        "type": "string",
                                        "example": "설명 1"
                                    },
                                    "data_json": {
                                        "description": "{}",
                                        "type": "string",
                                        "example": "{}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "결과값",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result_code": {
                                            "description": "성공:0, 실패:-1",
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "result_message": {
                                            "description": "성공:EMPTY, 실패:에러메세지(데이터 포맷 미 일치",
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "result_data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "이름",
                                                        "type": "string",
                                                        "example": "이름 1"
                                                    },
                                                    "description": {
                                                        "description": "설명",
                                                        "type": "string",
                                                        "example": "설명 1"
                                                    },
                                                    "data_json": {
                                                        "description": "JSON 문자열",
                                                        "type": "string",
                                                        "example": "{}"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api-key": {
                "type": "apiKey",
                "description": "Fast API Key",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    }
}